Support getting node version from nvm
Before this, the 'runtimeVersion' field in a launch configuration
could only contain a precise version of node to be used. You could
not defer the choice of version to the .nvmrc file in the project.

Now, by specifying `'runtimeVersion': 'nvm'`, we use `nvm exec` to
determine which version to use, which will pick up a version in
.nvmrc, if it exists, else use the version that nvm is currently
set to use.